feat: cargo deps -> update to last version (use u64 everywhere instead of usize bc doing `self.0.buffer_position().try_into().expect("XML position does not fit in usize")` would throw on 32 systems)

+++ src/types.rs
@@ -230,7 +230,7 @@ macro_rules! mk_id {
         fn from_str(s: &str) -> Result<Self, Self::Err> { <$ty>::from_str(s).map($id) }
       }
       impl $crate::parser::FromStrPos for $id {
-        fn to_err(_: Self::Err, pos: usize) -> $crate::parser::ParseError {
+        fn to_err(_: Self::Err, pos: u64) -> $crate::parser::ParseError {
           $crate::parser::ParseError::BadInteger(pos)
         }
       }
